move REST controllers to includes/API
