+++ includes/API/Site/Controllers/OnboardingController.php
@@ -8,13 +8,14 @@
  *
  * Onboarding data is stored locally in WordPress options.
  *
- * @package SnapchatForWooCommerce\Admin\Settings
+ * @package SnapchatForWooCommerce\API\Site\Controllers
  */
 
-namespace SnapchatForWooCommerce\Admin\Settings;
+namespace SnapchatForWooCommerce\API\Site\Controllers;
 
 use WP_REST_Response;
 use WP_REST_Request;
+use SnapchatForWooCommerce\Config;
 use SnapchatForWooCommerce\Utils\Storage\Options;
 use SnapchatForWooCommerce\Utils\Storage\OptionDefaults;
 
@@ -24,22 +25,12 @@ use SnapchatForWooCommerce\Utils\Storage\OptionDefaults;
  * @since 0.1.0
  */
 class OnboardingController extends SettingsBaseController {
-
-	/**
-	 * Class constructor.
-	 *
-	 * @since 0.1.0
-	 */
-	public function __construct() {
-		$this->namespace = 'wc/sfw/snapchat';
-	}
-
 	/**
 	 * Registers REST API routes.
 	 *
 	 * Provides GET and POST methods for onboarding state:
-	 * - GET  /wp-json/wc/sfw/snapchat/setup
-	 * - POST /wp-json/wc/sfw/snapchat/setup
+	 * - GET  /setup
+	 * - POST /setup
 	 *
 	 * @since 0.1.0
 	 *
@@ -47,18 +38,13 @@ class OnboardingController extends SettingsBaseController {
 	 */
 	public function register_routes(): void {
 		register_rest_route(
-			$this->namespace,
+			Config::REST_NAMESPACE . '/snapchat',
 			'/setup',
 			array(
 				array(
 					'methods'             => 'GET',
 					'permission_callback' => array( $this, 'permissions_check' ),
 					'callback'            => array( $this, 'get_setup_state' ),
-				),
-				array(
-					'methods'             => 'POST',
-					'permission_callback' => array( $this, 'permissions_check' ),
-					'callback'            => array( $this, 'set_setup_state' ),
 				),
 				'schema' => array( $this, 'setup_state_schema' ),
 			)
@@ -73,38 +59,6 @@ class OnboardingController extends SettingsBaseController {
 	 * @return WP_REST_Response
 	 */
 	public function get_setup_state(): WP_REST_Response {
-		return rest_ensure_response(
-			array(
-				'status' => Options::get( OptionDefaults::ONBOARDING_STATUS ),
-				'step'   => Options::get( OptionDefaults::ONBOARDING_STEP ),
-			)
-		);
-	}
-
-	/**
-	 * Sets the onboarding setup state.
-	 *
-	 * Accepts `status` and `step` fields in the request body and stores them
-	 * in WordPress options. Fields are optional, and only those provided
-	 * will be updated.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @param WP_REST_Request $request The REST request containing onboarding state data.
-	 * @return WP_REST_Response
-	 */
-	public function set_setup_state( WP_REST_Request $request ): WP_REST_Response {
-		$status = $request['status'] ?? '';
-		$step   = $request['step'] ?? '';
-
-		if ( $status ) {
-			Options::set( OptionDefaults::ONBOARDING_STATUS, $status );
-		}
-
-		if ( $step ) {
-			Options::set( OptionDefaults::ONBOARDING_STEP, $step );
-		}
-
 		return rest_ensure_response(
 			array(
 				'status' => Options::get( OptionDefaults::ONBOARDING_STATUS ),
