+++ includes/API/Site/Controllers/SettingsBaseController.php
@@ -7,13 +7,14 @@
  * and response formatting, and should be extended by all plugin-specific
  * settings controllers instead of using WC_REST_Controller directly.
  *
- * @package SnapchatForWooCommerce\Admin\Settings
+ * @package SnapchatForWooCommerce\API\Site\Controllers
  */
 
-namespace SnapchatForWooCommerce\Admin\Settings;
+namespace SnapchatForWooCommerce\API\Site\Controllers;
 
 use WC_REST_Controller;
-use WP_REST_Response;
+use SnapchatForWooCommerce\Config;
+use SnapchatForWooCommerce\API\AdminPermissionsTrait;
 
 /**
  * Abstract base controller for Ad Partner Settings REST API.
@@ -24,19 +25,5 @@ use WP_REST_Response;
  * @since 0.1.0
  */
 class SettingsBaseController extends WC_REST_Controller {
-
-	/**
-	 * Checks whether the current user has permission to access the settings endpoint.
-	 *
-	 * By default, this returns true for users with the `manage_woocommerce` capability.
-	 * This ensures that only store admins and authorized roles can modify settings
-	 * via the API.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @return bool True if the user has permission; false otherwise.
-	 */
-	public function permissions_check(): bool {
-		return current_user_can( 'manage_woocommerce' );
-	}
+	use AdminPermissionsTrait;
 }
